Move gamedata reading to be in the same thread as overlay
